Refactor user update handler to use string for AllowedContest and remove unnecessary user ID conversion
Enhance database connection handling by enabling pgcrypto extension during migration

Standardize UUID usage across models for primary keys and foreign keys

Update middleware authentication to handle string-based user IDs and roles

Refactor response utility to ensure consistent JSON response formatting

Implement CORS and logging middleware for improved request handling

Add preflight handling for OPTIONS requests in middleware

Ensure proper context management in middleware for user and engine data

+++ handlers/contest/create_contest.go
@@ -37,7 +37,7 @@ func (h *Handler) CreateContest(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	reqContest.Id = int64(newContest.ID)
+	reqContest.Id = newContest.ID
 	reqContest.CreatedAt = newContest.CreatedAt
 	reqContest.StartTime = newContest.StartTime
